Move the process browser to NewTools
Some weeks ago we discussed about the fact that we wanted all the new tools in the same repo. At that time the process browser was been developed, now that it's done I'm migrating it. 

I'm also making it the default one and moving it to the right place in the menu since it is stable enough to replace the previous one.