Improve folder selection in the File Browser
The method returning the selected entries of the file browser has some problems. 
First the code suggest that we are filtering folders to not return them, but an #and: was used instead of a #or: so this code is not effective. And this method is used also in the #OpenFolder variant of the file browser so the code should not be here.
Second, in a traditional file browser, if we allow to select a folder and we press confirm with no folder selected, then we get the current folder. I'm adapting NewTools FileBrowser to act in the same way.

Fixes #1229